+++ CHANGELOG.md
@@ -10,9 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - IMDb ratings
-  - IMDb vote count & rating to the engine
+  - IMDb vote count & rating to the rule engine
   - New task to refresh IMDb ratings daily (can be disabled if the user does not want IMDb)
     - This task daily will grab IMDb official rating data sets and update the database with the current ratings (these data sets are only refreshed once every 24 hours so running this task more than that is fruitless). There are guards that check the hash/304 not modified header that will prevent running if nothing has changed
+- AniList anime ratings
+  - Vote count, popularity, and favorites to the rule engine
+  - New task to refresh AniList ratings daily (can be disabled if the user does not want AniList)
 - Candidates view now shows date added metadata
 
 ### Changed